Ajustes nos pacotes
