+++ oni_google_backup_helper.egg-info/SOURCES.txt
@@ -1,0 +1,7 @@
+README.md
+setup.py
+oni_google_backup_helper.egg-info/PKG-INFO
+oni_google_backup_helper.egg-info/SOURCES.txt
+oni_google_backup_helper.egg-info/dependency_links.txt
+oni_google_backup_helper.egg-info/requires.txt
+oni_google_backup_helper.egg-info/top_level.txt
