+++ oni_google_backup_helper.egg-info/requires.txt
@@ -1,0 +1,30 @@
+cachetools
+certifi
+charset-normalizer
+google-api-core
+google-auth
+google-cloud-bigquery
+google-cloud-bigquery-datatransfer
+google-cloud-bigquery-storage
+google-cloud-core
+google-cloud-storage
+google-crc32c
+google-resumable-media
+googleapis-common-protos
+grpcio
+grpcio-status
+idna
+numpy
+packaging
+proto-plus
+protobuf
+pyarrow
+pyasn1
+pyasn1-modules
+pyparsing
+python-dateutil
+pytz
+requests
+rsa
+six
+urllib3
