+++ setup.py
@@ -17,7 +17,7 @@ setup(
     # Packages to include in the distribution:
     packages=find_packages(','),
     # Project version number:
-    version='0.1',
+    version='1.0',
     # List a license for the project, eg. MIT License
     license='',
     # Short description of your library:
@@ -37,36 +37,36 @@ setup(
     keywords=['onimusic'],
     # List project dependencies:
     install_requires=[
-        '<cachetools>',
-        '<certifi>',
-        '<charset-normalizer>',
-        '<google-api-core>',
-        '<google-auth>',
-        '<google-cloud-bigquery>',
-        '<google-cloud-bigquery-datatransfer>',
-        '<google-cloud-bigquery-storage>',
-        '<google-cloud-core>',
-        '<google-cloud-storage>',
-        '<google-crc32c>',
-        '<google-resumable-media>',
-        '<googleapis-common-protos>',
-        '<grpcio>',
-        '<grpcio-status>',
-        '<idna>',
-        '<numpy>',
-        '<packaging>',
-        '<proto-plus>',
-        '<protobuf>',
-        '<pyarrow>',
-        '<pyasn1>',
-        '<pyasn1-modules>',
-        '<pyparsing>',
-        '<python-dateutil>',
-        '<pytz>',
-        '<requests>',
-        '<rsa>',
-        '<six>',
-        '<urllib3>',
+        'cachetools',
+        'certifi',
+        'charset-normalizer',
+        'google-api-core',
+        'google-auth',
+        'google-cloud-bigquery',
+        'google-cloud-bigquery-datatransfer',
+        'google-cloud-bigquery-storage',
+        'google-cloud-core',
+        'google-cloud-storage',
+        'google-crc32c',
+        'google-resumable-media',
+        'googleapis-common-protos',
+        'grpcio',
+        'grpcio-status',
+        'idna',
+        'numpy',
+        'packaging',
+        'proto-plus',
+        'protobuf',
+        'pyarrow',
+        'pyasn1',
+        'pyasn1-modules',
+        'pyparsing',
+        'python-dateutil',
+        'pytz',
+        'requests',
+        'rsa',
+        'six',
+        'urllib3',
     ],
     # https://pypi.org/classifiers/
     classifiers=["Private :: Do Not Upload"]
